Refactored management of build workers
Moved practically all of the Jenkins build logic into java; PackageGraphExecution now manages the submission to the Jenkins queue
of package builds, along with allocation of workspace, nodes, etc. In general I like the idea of composing complex workflows with
a DSL built on Groovy, but in practice, it's quite opaque, and options for reusing code are unsatisfactory:

- If you load the "flow script" from SCM, it *must* run in the sandbox, leading to hours of wasted time wading through
  script approvals and what appear to bugs in the sandbox mechanism, AND, you are basically limited to loading from a single file,
  which makes it difficult to reuse blocks of logic

- There was something weird going on with the parallel() step and access to our BuildQueue instance; in principle, workers should
  have unfettered access to the BuildQueue instance, which ensures thread safety through the use of a SynchronousQueue, but in
  practice it seems that some sort of serialization happens, preventing package builds from running in parallel. Most annoying.

When we get rid of all the groovy crap and launch our own tasks, everything seems to work very well. Jobs survive Jenkins restarts,
and to date we are able to achieve 80% of CPU usage on a 16-core VM. Mazeltof.