Added ability to disable package versions
Some very old package versions simply won't work with Renjin -- or GNU R.
For example, old versions of the zoo package used fixInNamespace() to
repalce as.Date.numeric. This will not work anymore, either in Renjin or in GNU R.

So now through the user interface, admins (us) can mark project versions as disabled,
in which case they will not be build or tested, and won't be included in
dependency resolution.