Implemented performance regression detection using changepoint detection
This is an initial working version that uses the changepoint package to
find changes in performance within benchmarks.

Right now, it does not distinguish between different JVM versions,
however, so the results are still not super useful.

I think the next step is to further segment the analysis not only by the
benchmark machine, but also the JVM version.